Create multi_tine.py
Changed tine.py so there is the option to play two tines at once. I added the option to change the Density and the Rigidity since those seem optimal for changing the tine sounds. It is hard coded, not through the command prompt however which I can change that later if needed :)